Revert "Feature/container modal"
