--- .prettierrc.json
@@ -1,13 +1,0 @@
-{
-  "arrowParens": "always",
-  "bracketSpacing": true,
-  "jsxBracketSameLine": false,
-  "jsxSingleQuote": false,
-  "quoteProps": "as-needed",
-  "singleQuote": true,
-  "semi": true,
-  "printWidth": 100,
-  "useTabs": false,
-  "tabWidth": 2,
-  "trailingComma": "es5"
-}
