--- components/input.jsx
@@ -1,5 +1,0 @@
-import { Input } from '@rneui/themed';
-
-export default (props) => {
-  return <Input placeholder={props.placeholder} secureTextEntry={props.secure} {...props} />;
-};
