--- components/logo.jsx
@@ -1,27 +1,0 @@
-import Svg, { Path } from 'react-native-svg';
-import { View } from 'react-native';
-
-const Logo = (props) => (
-  <Svg width={props.width} height={props.height} xmlns="http://www.w3.org/2000/svg" {...props}>
-    <Path
-      d="M92.0889 0C18.0889 0 0 18.8274 0 92.5C0 166.173 18.0889 185 92.0889 185C166.089 185 185 166.991 185 92.5C185 18.8274 166.911 0 92.0889 0Z"
-      fill="#6C0263"
-    />
-    <Path
-      d="M140.379 125.18C140.697 125.18 141.022 125.199 141.353 125.237C141.685 125.275 141.978 125.371 142.232 125.524C142.5 125.677 142.717 125.912 142.882 126.231C143.048 126.549 143.131 126.989 143.131 127.549C143.131 128.072 143.048 128.486 142.882 128.791C142.717 129.084 142.5 129.314 142.232 129.479C141.978 129.632 141.685 129.734 141.353 129.785C141.022 129.823 140.697 129.842 140.379 129.842H139.232V139.97C139.232 140.53 139.201 141.072 139.137 141.594C139.086 142.104 138.952 142.562 138.736 142.97C138.532 143.365 138.226 143.683 137.818 143.925C137.423 144.167 136.888 144.288 136.213 144.288C135.564 144.288 135.048 144.167 134.665 143.925C134.296 143.683 134.016 143.365 133.825 142.97C133.634 142.562 133.513 142.104 133.462 141.594C133.411 141.072 133.385 140.53 133.385 139.97V129.842H132.315C131.958 129.842 131.614 129.823 131.283 129.785C130.965 129.734 130.678 129.632 130.423 129.479C130.181 129.314 129.984 129.084 129.831 128.791C129.678 128.486 129.602 128.072 129.602 127.549C129.602 126.989 129.678 126.549 129.831 126.231C129.984 125.912 130.181 125.677 130.423 125.524C130.678 125.371 130.965 125.275 131.283 125.237C131.614 125.199 131.958 125.18 132.315 125.18H133.385C133.385 124.059 133.436 123.027 133.538 122.084C133.64 121.141 133.888 120.339 134.283 119.676C134.678 119.001 135.264 118.479 136.041 118.11C136.818 117.727 137.882 117.536 139.232 117.536C139.806 117.536 140.347 117.587 140.857 117.689C141.366 117.791 141.812 117.95 142.194 118.167C142.576 118.383 142.876 118.67 143.092 119.027C143.322 119.371 143.436 119.791 143.436 120.288C143.436 120.747 143.354 121.129 143.188 121.434C143.035 121.74 142.844 121.989 142.615 122.18C142.385 122.371 142.131 122.505 141.85 122.581C141.583 122.657 141.334 122.696 141.105 122.696C140.914 122.696 140.736 122.651 140.57 122.562C140.404 122.473 140.213 122.428 139.997 122.428C139.818 122.428 139.678 122.485 139.576 122.6C139.487 122.715 139.411 122.868 139.347 123.059C139.296 123.25 139.264 123.46 139.252 123.689C139.239 123.919 139.232 124.148 139.232 124.377V125.18H140.379ZM148.328 151.932C146.876 151.932 145.844 151.722 145.233 151.301C144.621 150.894 144.315 150.289 144.315 149.486C144.315 149.091 144.366 148.76 144.468 148.492C144.583 148.225 144.736 148.008 144.927 147.843C145.118 147.677 145.334 147.556 145.576 147.48C145.831 147.416 146.099 147.384 146.379 147.384C146.57 147.384 146.748 147.39 146.914 147.403C147.08 147.416 147.22 147.422 147.334 147.422C147.577 147.422 147.806 147.276 148.022 146.983C148.252 146.703 148.468 146.346 148.672 145.913C148.889 145.48 149.105 145.015 149.322 144.518C149.538 144.021 149.755 143.562 149.971 143.142L149.666 142.378C149.296 141.448 148.908 140.473 148.5 139.454C148.092 138.422 147.685 137.384 147.277 136.339C146.882 135.282 146.487 134.231 146.092 133.186C145.71 132.142 145.36 131.129 145.041 130.148C144.901 129.74 144.78 129.339 144.678 128.944C144.589 128.549 144.545 128.161 144.545 127.779C144.545 127.409 144.621 127.065 144.774 126.747C144.927 126.428 145.137 126.154 145.405 125.925C145.672 125.696 145.978 125.517 146.322 125.39C146.678 125.25 147.054 125.18 147.449 125.18C147.806 125.18 148.131 125.256 148.424 125.409C148.729 125.562 148.997 125.772 149.226 126.04C149.468 126.307 149.678 126.613 149.857 126.957C150.035 127.301 150.188 127.664 150.315 128.046L152.876 136.11L155.322 128.046C155.437 127.689 155.583 127.333 155.761 126.976C155.94 126.619 156.15 126.307 156.392 126.04C156.634 125.759 156.908 125.536 157.214 125.371C157.519 125.192 157.857 125.103 158.226 125.103C158.596 125.103 158.946 125.173 159.277 125.314C159.621 125.454 159.914 125.651 160.156 125.906C160.411 126.148 160.609 126.441 160.749 126.785C160.902 127.129 160.978 127.498 160.978 127.893C160.978 128.275 160.927 128.658 160.825 129.04C160.736 129.409 160.621 129.779 160.481 130.148L155.895 142.301C155.628 142.989 155.335 143.702 155.016 144.441C154.71 145.193 154.373 145.925 154.003 146.639C153.634 147.352 153.239 148.027 152.819 148.664C152.398 149.301 151.952 149.862 151.481 150.346C151.01 150.83 150.507 151.212 149.971 151.492C149.449 151.785 148.901 151.932 148.328 151.932Z"
-      fill="white"
-    />
-    <Path
-      d="M92.9581 40.9935C97.184 40.9935 101.149 41.5395 104.852 42.6316C108.603 43.7237 111.856 45.3618 114.61 47.546C117.411 49.7302 119.596 52.4604 121.163 55.7367C122.777 58.9655 123.584 62.7403 123.584 67.0612C123.584 69.2929 123.275 71.3346 122.658 73.1864C122.088 74.9907 121.329 76.6289 120.379 78.1008C119.477 79.5728 118.456 80.8548 117.316 81.9469C116.224 83.039 115.156 83.9886 114.111 84.7958C113.067 85.5555 112.141 86.1728 111.334 86.6476C110.527 87.075 109.957 87.3361 109.624 87.4311C112.473 88.2383 114.871 89.2829 116.818 90.5649C118.765 91.7994 120.379 93.1764 121.661 94.6958C122.991 96.1678 124.011 97.711 124.724 99.3254C125.483 100.892 126.029 102.435 126.362 103.955C126.742 105.427 126.955 106.804 127.003 108.086C127.098 109.32 127.145 110.317 127.145 111.077C127.145 114.781 126.694 118.057 125.792 120.906C124.89 123.707 123.703 126.153 122.231 128.242C120.759 130.284 119.073 132.041 117.174 133.513C115.275 134.937 113.352 136.1 111.405 137.003C106.752 139.187 101.505 140.421 95.6646 140.706H72.3033C69.5968 140.706 67.5076 140.255 66.0357 139.353C64.5637 138.403 63.4954 137.169 62.8306 135.649C62.1659 134.082 61.7623 132.326 61.6198 130.379C61.5249 128.432 61.4774 126.414 61.4774 124.325V57.2324C61.4774 55.1906 61.5249 53.1964 61.6198 51.2496C61.7623 49.3028 62.1659 47.5697 62.8306 46.0503C63.4954 44.5309 64.5637 43.3201 66.0357 42.4179C67.5076 41.4683 69.5968 40.9935 72.3033 40.9935H92.9581ZM83.1293 97.1174V123.897H93.5279C95.7595 123.802 97.7538 123.28 99.5106 122.33C100.27 121.951 101.006 121.452 101.719 120.835C102.431 120.218 103.072 119.458 103.642 118.556C104.211 117.606 104.662 116.538 104.995 115.351C105.327 114.116 105.493 112.692 105.493 111.077C105.493 109.32 105.303 107.801 104.924 106.519C104.591 105.189 104.116 104.05 103.499 103.1C102.929 102.103 102.265 101.272 101.505 100.607C100.745 99.8951 99.9855 99.3254 99.2257 98.898C97.3739 97.9009 95.2847 97.3074 92.9581 97.1174H83.1293ZM83.1293 57.8022V82.0181H92.2459C94.2876 81.9231 96.1157 81.4483 97.7301 80.5936C98.3948 80.2613 99.0596 79.8102 99.7243 79.2404C100.389 78.6706 100.983 78.0058 101.505 77.2461C102.027 76.4389 102.431 75.4893 102.716 74.3972C103.048 73.3051 103.214 72.0468 103.214 70.6224C103.214 69.008 103.048 67.6072 102.716 66.4202C102.383 65.1856 101.932 64.141 101.362 63.2864C100.84 62.3842 100.223 61.6245 99.5106 61.0072C98.8459 60.3899 98.1574 59.8676 97.4452 59.4403C95.7358 58.5381 93.8128 57.9921 91.6761 57.8022H83.1293Z"
-      fill="white"
-    />
-    <Path
-      d="M74.4685 38.364C69.1824 42.15 61.0025 32.9975 61.0025 32.9975C61.0025 32.9975 67.0349 43.6881 61.7489 47.4742C56.4628 51.2602 48.283 42.1077 48.283 42.1077L60.776 59.5504L86.2152 41.33L73.7221 23.8874C73.7221 23.8874 79.7545 34.5779 74.4685 38.364Z"
-      fill="#F0E12B"
-      stroke="#F0E12B"
-      strokeWidth="0.508242"
-    />
-  </Svg>
-);
-
-export default Logo;
