--- components/searchbar.jsx
@@ -1,87 +1,0 @@
-import React from "react";
-import {StyleSheet, TextInput, View, Keyboard, Button} from "react-native";
-import {Feather, Entypo} from "@expo/vector-icons";
-
-const SearchBar = ({clicked, searchPhrase, setSearchPhrase, setCLicked}) => {
-	return (
-		<View style={styles.container}>
-			<View
-				style={
-					clicked
-						? styles.searchBar__clicked
-						: styles.searchBar__unclicked
-				}
-			>
-				{/* search Icon */}
-				<Feather
-					name="search"
-					size={20}
-					color="black"
-					style={{marginLeft: 1}}
-				/>
-				{/* Input field */}
-				<TextInput
-					style={styles.input}
-					placeholder="Search"
-					value={searchPhrase}
-					onChangeText={setSearchPhrase}
-					onFocus={() => {
-						setClicked(true);
-					}}
-				/>
-				{/* cross Icon, depending on whether the search bar is clicked or not */}
-				{clicked && (
-					<Entypo name="cross" size={20} color="black" style={{padding: 1}} onPress={() => {
-						setSearchPhrase("")
-					}}/>
-				)}
-			</View>
-			{/* cancel button, depending on whether the search bar is clicked or not */}
-			{clicked && (
-				<View>
-					<Button
-						title="Cancel"
-						onPress={() => {
-							Keyboard.dismiss();
-							setClicked(false);
-						}}
-					></Button>
-				</View>
-			)}
-		</View>
-	);
-};
-export default SearchBar;
-
-// styles
-const styles = StyleSheet.create({
-	container: {
-		margin: 15,
-		alignSelf:"center",
-		width: "50%",
-
-	},
-	searchBar__unclicked: {
-		padding: 10,
-		flexDirection: "row",
-		width: "95%",
-		backgroundColor: "#FFFFFF",
-		borderRadius: 15,
-		alignItems: "center",
-		boxShadow: "1px 3px 10px #9E9E9E"
-	},
-	searchBar__clicked: {
-		padding: 10,
-		flexDirection: "row",
-		width: "80%",
-		backgroundColor: "#D9DBDA",
-		borderRadius: 15,
-		alignItems: "center",
-		justifyContent: "space-evenly",
-	},
-	input: {
-		fontSize: 20,
-		marginLeft: 10,
-		width: "90%",
-	},
-});
