--- styles/theme.js
@@ -1,12 +1,0 @@
-import { createTheme } from '@rneui/themed';
-
-const theme = createTheme({
-  lightColors: {
-    primary: '#6c0263',
-  },
-  darkColors: {
-    primary: '#000',
-  },
-});
-
-export default theme;
