--- webpack.config.js
@@ -1,11 +1,0 @@
-const createExpoWebpackConfigAsync = require('@expo/webpack-config');
-
-module.exports = async function(env, argv) {
-    const config = await createExpoWebpackConfigAsync({
-        ...env,
-        babel: {
-            dangerouslyAddModulePathsToTranspile: ['@rneui']
-        }
-    }, argv);
-    return config;
-};
